Make Jest test actually compile to ESM module.
I knew that was too easy! I should've known when that "actual require"
method worked 😅.

`jest.config.ts` needed to be told to treat `.ts` extensions as ESM
modules. Then I needed to change quite a few things to ensure that we
dynamically import `db` so it can be mocked.